requirements.txt
torch
torchvision
transformers
Pillow
ultralytics
mediapipe
insightface
onnxruntime
faiss-cpu
opencv-python
streamlit
scikit-learn
numpy
watchdog
